requirements.txt
requests
beautifulsoup4
python-dateutil
selenium
selenium-stealth
webdriver-manager
pdfplumber
backports.zoneinfo; python_version < "3.9"
pillow
